refactor: relocate constants file
